+++ auth.ts
@@ -1,14 +1,12 @@
 import NextAuth from "next-auth"
-import MicrosoftEntraID from "next-auth/providers/microsoft-entra-id"
+import AzureAD from "next-auth/providers/azure-ad"
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
     providers: [
-        MicrosoftEntraID({
-            id: "azure-ad",
+        AzureAD({
             clientId: process.env.AUTH_MICROSOFT_ID,
             clientSecret: process.env.AUTH_MICROSOFT_SECRET,
-            issuer: `https://login.microsoftonline.com/${process.env.AUTH_MICROSOFT_TENANT_ID}/v2.0`,
-            authorization: { params: { scope: "openid profile email User.Read" } },
+            tenantId: process.env.AUTH_MICROSOFT_TENANT_ID,
         }),
     ],
     callbacks: {